Fix sb-refresh, and allow --update to be provided the :batch modifier When running with :batch, the program assumes that the cache has already been deleted by sb-refresh, and thus can use the re-generated cache from other programs.

+++ cpp/src/main.cpp
@@ -125,7 +125,12 @@ int main(int argc, char* argv[]) {
   }
 
   if (arg::at("update") >= "cache") {
-    std::filesystem::remove_all(std::filesystem::path(data) / "sb" / "cache");
+
+    // So sb-refresh can reuse caches populated by other sandboxes.
+    // We don't need to repeat the work.
+    if (arg::mod("update") != "batch")
+      std::filesystem::remove_all(std::filesystem::path(data) / "sb" / "cache");
+
     std::filesystem::remove_all(std::filesystem::path(data) / "sb" / program / "cache");
     std::filesystem::remove_all(std::filesystem::path(libraries::hash_sof(program, arg::hash)));
   }
